Make cli async
